pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autodev"
version = "0.1.0"
description = "Autonomous multi-agent coding system with strict git hygiene and monitoring"
requires-python = ">=3.11"

dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "gitpython>=3.1.43",
  "redis>=5.0.7",
  "rq>=1.16.2",
  "anthropic>=0.34.1",
  "openai>=1.40.0",
  "google-genai>=0.3.0",          # Official GenAI SDK for Gemini API
  "httpx>=0.27.0",
  "pyyaml>=6.0.1",
  "typer>=0.12.3",
  "rich>=13.7.1",
  "prometheus-client>=0.20.0",    # Prometheus Python client
  "apscheduler>=3.10.4",
  "libtmux>=0.28.0",              # programmatic tmux control
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",                # test runner
  "mypy>=1.10.0",                 # static typing
  "ruff>=0.5.3",                  # linter/formatter
  "pre-commit>=3.7.1",
  "commitizen>=3.27.0",           # Conventional Commits automation
]

[tool.ruff]
line-length = 100
target-version = "py311"
lint.select = ["E", "F", "I", "UP", "B", "Q", "RUF", "W"]
format.quote-style = "double"

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
tag_format = "v$version"
update_changelog_on_bump = true